requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
dnspython==2.7.0
fastapi==0.115.12
h11==0.16.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
motor==3.7.0
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pymongo==4.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
rich==14.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
